Moved the get job script to script folder
